print logo

+++ src/driver/mod.rs
@@ -4,7 +4,7 @@ pub mod uart;
 pub mod gpio;
 
 /// Read something from mmio address
-pub fn mmio_read(addr: u32) -> u32 {
+pub fn mmio_read(addr: usize) -> u32 {
     unsafe{
         let val = read(addr as *const u32);
         return val
